On The Map App
An app with a map that shows information posted by other students. The
map will contain pins that show the location where other students have
reported studying. By tapping on the pin users can see a URL for
something the student finds interesting. The user will be able to add
their own data by posting a string that can be reverse geocoded to a
location, and a URL.